Use intp

+++ pandas/tests/indexes/ranges/test_range.py
@@ -614,7 +614,7 @@ def test_reindex_returns_rangeindex():
     expected = RangeIndex(1, 4, name="foo")
     tm.assert_index_equal(result, expected, exact=True)
 
-    expected_indexer = np.array([1, -1, -1])
+    expected_indexer = np.array([1, -1, -1], dtype=np.intp)
     tm.assert_numpy_array_equal(result_indexer, expected_indexer)
